Scaling Feast 1.9 (#132)
* We fix #114 and add a lower bound to max hunger

* Actually make Bloated Overflow toggleable, closes #121

* Fix #118 no more NPE from getMaxHunger

* Fix adding material stats before conarm does preinit, closes #128

* Prevent FakePlayer from getting Bloated Overlflow, closing #122

* Fix random crash for SOLCarrot integration, closes #117

* Complete HUD drawing rework, tweaks for food values display for edible blocks

* Fix Exhaustium material not being coloured correctly.

* Exhaustion Scaling feature in #111

* The Couch Potato has arrived

* fix recipe

* new logo, tweaks to modifiers and starvation

* Extra effects for Molten Exhaustion

* Feature Blacklist/Whitelist for #126

* clean up debug messages

+++ src/main/java/yeelp/scalingfeast/util/Colour.java
@@ -41,8 +41,8 @@ public class Colour {
 	 * 
 	 * @return the red component of this Colour
 	 */
-	public int getR() {
-		return (this.r < 0 ? this.r + 256 : this.r);
+	public byte getR() {
+		return (byte) (this.r < 0 ? this.r + 256 : this.r);
 	}
 
 	/**
@@ -50,8 +50,8 @@ public class Colour {
 	 * 
 	 * @return the green component of this Colour
 	 */
-	public int getG() {
-		return (this.g < 0 ? this.g + 256 : this.g);
+	public byte getG() {
+		return (byte) (this.g < 0 ? this.g + 256 : this.g);
 	}
 
 	/**
@@ -59,7 +59,7 @@ public class Colour {
 	 * 
 	 * @return the blue component of this Colour
 	 */
-	public int getB() {
-		return (this.b < 0 ? this.b + 256 : this.b);
+	public byte getB() {
+		return (byte) (this.b < 0 ? this.b + 256 : this.b);
 	}
 }
